agentic_system/requirements.txt
together
numpy
scikit-learn
python-dotenv
streamlit
streamlit-ace
psycopg2-binary
sqlalchemy
PyPDF2
python-docx
odfpy
Pillow
pytesseract
requests




